Add demo with ImageBitmapRenderingContext
The main thread controls a visible canvas that has a
`bitmaprenderer` context (i.e. `ImageBitmapRenderingContext`).

The web worker creates a `OffscreenCanvas`, renders into it,
converts the content of the offscreen canvas into a bitmap, then
sends the bitmap to the main thread with a zero-copy operation.

The main thread takes the bitmap and calls
`transferFromImageBitmap` on the `ImageBitmapRenderingContext`
to update its visible canvas.