Add demo with a transfer of ownership of a canvas
In this demo the main thread transfers the ownership of a canvas
to a OffscreenCanvas managed by a web worker. The web worker
renders into this OffscreenCanvas in a render loop, implemented
with requestAnimationFrame in the web worker itself. The main
thread can send a message to the web worker to start/stop the
render loop. The visible, onscreen canvas gets updated
automatically and asynchronously.